[convert-to-use-perl] added simple converter
I can write HTML, but I don't want to manually add all the crazy stuff
use.perl forces me to do to make the posts come out right. Better to
let a script do that. This script puts <p>...</p> around comment
paragraphs, and escapes the '<' and '>' characters in code paragraphs.
It also turns all the spaces in the latter into &nbsp; so that when
the user copies and pastes, code doesn't get wrapped.